refactor: Remove override and add extend
