+++ config/application.rb
@@ -46,20 +46,27 @@ module DevelopmentApp
     end
 
     config.after_initialize do
+      # Controllers
       require "extends/controllers/decidim/devise/sessions_controller_extends"
       require "extends/controllers/decidim/editor_images_controller_extends"
-      require "extends/services/decidim/iframe_disabler_extends"
-      require "extends/helpers/decidim/icon_helper_extends"
-      require "extends/commands/decidim/initiatives/admin/update_initiative_answer_extends"
+      require "extends/controllers/decidim/proposals/proposals_controller_extends"
+      require "extends/controllers/decidim/newsletters_controller_extends"
       require "extends/controllers/decidim/initiatives/committee_requests_controller_extends"
+      # Models
       require "extends/models/decidim/budgets/project_extends"
       require "extends/models/decidim/authorization_extends"
-      require "extends/forms/decidim/initiatives/admin/initiative_form_extends"
-      require "extends/commands/decidim/budgets/admin/import_proposals_to_budgets_extends"
-      require "extends/controllers/decidim/newsletters_controller_extends"
-      require "extends/commands/decidim/admin/destroy_participatory_space_private_user_extends"
-      require "extends/controllers/decidim/proposals/proposals_controller_extends"
+      # Services
+      require "extends/services/decidim/iframe_disabler_extends"
+      # Helpers
+      require "extends/helpers/decidim/icon_helper_extends"
+      # Forms
       require "extends/forms/decidim/initiatives/initiative_form_extends"
+      require "extends/forms/decidim/initiatives/admin/initiative_form_extends"
+      # Commands
+      require "extends/commands/decidim/initiatives/admin/update_initiative_answer_extends"
+      require "extends/commands/decidim/budgets/admin/import_proposals_to_budgets_extends"
+      require "extends/commands/decidim/admin/destroy_participatory_space_private_user_extends"
+      require "extends/commands/decidim/admin/create_attachment_extends"
 
       Decidim::GraphiQL::Rails.config.tap do |config|
         config.initial_query = "{\n  deployment {\n    version\n    branch\n    remote\n    upToDate\n    currentCommit\n    latestCommit\n    locallyModified\n  }\n}".html_safe
